test_quick.py
"""Quick test of OpenAI connection"""
import os
from dotenv import load_dotenv

load_dotenv()

print("Testing OpenAI Connection...")
print("-" * 60)

# Check environment
api_key = os.getenv('OPENAI_API_KEY')
print(f"1. API Key loaded: {bool(api_key)}")
if api_key:
    print(f"   Starts with: {api_key[:15]}...")

# Test import
try:
    from openai import OpenAI
    print("2. OpenAI package imported: ✅")
except Exception as e:
    print(f"2. OpenAI package import failed: ❌ {e}")
    exit(1)

# Test client creation
try:
    client = OpenAI(api_key=api_key)
    print("3. OpenAI client created: ✅")
except Exception as e:
    print(f"3. Client creation failed: ❌ {e}")
    exit(1)

# Test simple API call
try:
    print("4. Testing API call...")
    response = client.chat.completions.create(
        model="gpt-4o-mini",
        messages=[
            {"role": "system", "content": "You are a helpful assistant."},
            {"role": "user", "content": "Say 'Hello' in JSON format: {\"message\": \"Hello\"}"}
        ],
        temperature=0.2,
        max_tokens=50
    )
    
    result = response.choices[0].message.content
    print(f"   API Response: {result[:100]}")
    print("4. API call successful: ✅")
    
except Exception as e:
    print(f"4. API call failed: ❌ {e}")
    exit(1)

# Test our generate_requirements function
try:
    print("5. Testing generate_requirements function...")
    import config
    import importlib
    importlib.reload(config)
    
    from app.services.ai_client import generate_requirements
    
    # Simple test
    reqs = generate_requirements("Test requirement", {"key": "value"})
    print(f"   Generated {len(reqs)} requirements")
    if reqs:
        print(f"   First requirement: {reqs[0]['title']}")
        print(f"   Status: {reqs[0]['status']}")
    print("5. generate_requirements works: ✅")
    
except Exception as e:
    print(f"5. generate_requirements failed: ❌ {e}")
    import traceback
    traceback.print_exc()
    exit(1)

print("-" * 60)
print("🎉 ALL TESTS PASSED!")
print()
print("The AI Agent is fully functional:")
print("  ✅ OpenAI API connection working")
print("  ✅ generate_requirements function working")
print("  ✅ JSON parsing working")
print("  ✅ Status set to 'Offen'")
