requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.11.12
click==8.1.8
colorama==0.4.6
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
openai==2.8.1
openpyxl==3.1.2
packaging==25.0
pillow==11.3.0
pydantic==2.12.4
pydantic_core==2.41.5
reportlab==4.0.7
sniffio==1.3.1
SQLAlchemy==2.0.43
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.3
WTForms==3.2.1
python-dotenv==1.0.1
zipp==3.23.0
